Handle incoming SDI SOAP requests
Our SOAP server needs to be able to parse different types of incoming
requests. Here we prepare a structure that accepts the different types
of messages that SDI might send us and we parse the message.

We still need to decide how we want to pass in the handling of the message.